Move $raffleId to subfolder
